# avoid writing and reparsing the artifact usage file once per artifact file in the project deps