> When in doubt, do things in bulk

The timings below will vary depending on the size of the artifact usage log file but I guess my computer is somewhat representative:

```
julia> ENV["JULIA_PKG_PRECOMPILE_AUTO"] = 0
0

julia> Pkg.activate(; temp=true)

julia> @btime Pkg.add("Plots")
# master
  618.572 ms (4700447 allocations: 339.36 MiB)

# PR
  428.434 ms (3376261 allocations: 268.50 MiB)
``` 

Looking at a TimerOutput profile we can see how the runtime for `download_artifacts` changes:

Master:

```
───────────────────────────────────────────────────────────────────────────────────────────────
                                                      Time                    Allocations      
                                             ───────────────────────   ────────────────────────
              Tot / % measured:                   829ms /  99.4%            340MiB /  99.3%    

Section                              ncalls     time    %tot     avg     alloc    %tot      avg
───────────────────────────────────────────────────────────────────────────────────────────────
add                                       1    825ms  100.0%   825ms    338MiB  100.0%   338MiB
  add                                     1    825ms  100.0%   825ms    338MiB  100.0%   338MiB
...
    download_artifacts                    1    186ms   22.6%   186ms   75.8MiB   22.4%  75.8MiB
...
───────────────────────────────────────────────────────────────────────────────────────────────
```


PR:
```
───────────────────────────────────────────────────────────────────────────────────────────────
                                                      Time                    Allocations      
                                             ───────────────────────   ────────────────────────
              Tot / % measured:                   507ms /  99.6%            269MiB /  99.4%    

Section                              ncalls     time    %tot     avg     alloc    %tot      avg
───────────────────────────────────────────────────────────────────────────────────────────────
add                                       1    505ms  100.0%   505ms    267MiB  100.0%   267MiB
  add                                     1    505ms  100.0%   505ms    267MiB  100.0%   267MiB
...
    download_artifacts                    1   7.46ms    1.3%  7.46ms   7.38MiB    2.7%  7.38MiB
...    
───────────────────────────────────────────────────────────────────────────────────────────────
```

186ms  -> 5.68 ms. 75 MB > 7.38 MB.