@IanButterworth 2024-12-17
Sounds good. Any benefit of unique!-ing the list too?